Make fact_uint64() faster
In `fact_uint64()`, speed up the calculation of the result by halving
the number of multiplications.

This is not my original idea, but the basic idea is to rearrange the
factors and use some shortcuts.

For example, for even numbers > 1:

    10! = 10.9.8.7.6.5.4.3.2.1
        = (10.1).(9.2).(8.3).(7.4).(6.5)
        = 10.18.24.28.30
        = 3628800

The speed up is to use the third line, which has less than half the
number of multiplications as the first line, but the factors to be
multiplied still follow a simple pattern that involves only additive
operations.

For odd numbers > 1, calculate it in the same way as the next even
number but drop the first factor:

    9!  = 9.8.7.6.5.4.3.2.1
        = (9.2).(8.3).(7.4).(6.5)
        = 18.24.28.30
        = 362880